split components into multiple files
