+++ ui/pages/confirmations/components/confirm/info/approve/spending-cap/spending-cap.tsx
@@ -1,0 +1,123 @@
+import { TransactionMeta } from '@metamask/transaction-controller';
+import React from 'react';
+import { useSelector } from 'react-redux';
+import {
+  ConfirmInfoRow,
+  ConfirmInfoRowDivider,
+  ConfirmInfoRowText,
+} from '../../../../../../../components/app/confirm/info/row';
+import { ConfirmInfoSection } from '../../../../../../../components/app/confirm/info/row/section';
+import Tooltip from '../../../../../../../components/ui/tooltip';
+import { useI18nContext } from '../../../../../../../hooks/useI18nContext';
+import { currentConfirmationSelector } from '../../../../../../../selectors';
+import { useAssetDetails } from '../../../../../hooks/useAssetDetails';
+import { Container } from '../../shared/transaction-data/transaction-data';
+import { getAccountBalance } from '../edit-spending-cap-modal/edit-spending-cap-modal';
+import {
+  UNLIMITED_MSG,
+  useApproveTokenSimulation,
+} from '../hooks/use-approve-token-simulation';
+
+const SpendingCapGroup = ({
+  tokenSymbol,
+  decimals,
+  setIsOpenEditSpendingCapModal,
+  customSpendingCap,
+}: {
+  tokenSymbol: string;
+  decimals: string;
+  setIsOpenEditSpendingCapModal: (newValue: boolean) => void;
+  customSpendingCap: string;
+}) => {
+  const t = useI18nContext();
+
+  const transactionMeta = useSelector(
+    currentConfirmationSelector,
+  ) as TransactionMeta;
+
+  const { spendingCap, formattedSpendingCap, value } =
+    useApproveTokenSimulation(transactionMeta, decimals);
+
+  const SpendingCapElement = (
+    <ConfirmInfoRowText
+      text={
+        spendingCap === UNLIMITED_MSG
+          ? `${t('unlimited')} ${tokenSymbol}`
+          : `${
+              customSpendingCap === ''
+                ? formattedSpendingCap
+                : customSpendingCap
+            } ${tokenSymbol}`
+      }
+      onEditClick={() => setIsOpenEditSpendingCapModal(true)}
+      editIconClassName="edit-spending-cap-icon"
+    />
+  );
+
+  if (!value) {
+    return null;
+  }
+
+  return (
+    <>
+      <ConfirmInfoRowDivider />
+
+      <ConfirmInfoRow
+        label={t('spendingCap')}
+        tooltip={t('spendingCapTooltipDesc')}
+      >
+        {spendingCap === UNLIMITED_MSG ? (
+          <Tooltip title={formattedSpendingCap}>{SpendingCapElement}</Tooltip>
+        ) : (
+          SpendingCapElement
+        )}
+      </ConfirmInfoRow>
+    </>
+  );
+};
+
+export const SpendingCap = ({
+  setIsOpenEditSpendingCapModal,
+  customSpendingCap,
+}: {
+  setIsOpenEditSpendingCapModal: (newValue: boolean) => void;
+  customSpendingCap: string;
+}) => {
+  const t = useI18nContext();
+
+  const transactionMeta = useSelector(
+    currentConfirmationSelector,
+  ) as TransactionMeta;
+
+  const { userBalance, tokenSymbol, decimals } = useAssetDetails(
+    transactionMeta.txParams.to,
+    transactionMeta.txParams.from,
+    transactionMeta.txParams.data,
+  );
+
+  const accountBalance = getAccountBalance(userBalance || '0', decimals || '0');
+
+  const { pending } = useApproveTokenSimulation(
+    transactionMeta,
+    decimals || '0',
+  );
+
+  if (pending) {
+    return <Container isLoading />;
+  }
+
+  return (
+    <ConfirmInfoSection>
+      <ConfirmInfoRow label={t('accountBalance')}>
+        <ConfirmInfoRowText text={`${accountBalance} ${tokenSymbol || ''}`} />
+      </ConfirmInfoRow>
+
+      <SpendingCapGroup
+        tokenSymbol={tokenSymbol || ''}
+        decimals={decimals || '0'}
+        setIsOpenEditSpendingCapModal={setIsOpenEditSpendingCapModal}
+        customSpendingCap={customSpendingCap}
+      />
+    </ConfirmInfoSection>
+  );
+};
